requirements.txt
Flask
Pyyaml